\name{glioma.cov}
\alias{glioma.cov}
\title{
Age and gender of the 85 patients with  gliomas
}
\description{
This data set consists of age and gender the 85 patients with gliomas.  
}
\usage{data(glioma)}

\format{a matrix containing age and gender of 85 samples}
\source{
Freije et al. (2004). Gene Expression Profiling of Gliomas Strongly Predicts Survival, Cancer Research, 64: 6503-6510. 
}
\keyword{datasets}

