\name{glioma}
\alias{glioma}
\title{
Gene expression and survival data of the patients with  gliomas
}
\description{
These data sets consist of gene expression and survival of the patients with gliomas.  Note that it contains a subset of the data published in Freije et al. (2004). 
}
\usage{data(glioma)}

\format{a matrix x containing 200 genes and 85 samples; a matrix y containing survival, censoring status, age, gender, and stage of 85 samples}
\source{
Freije et al. (2004). Gene Expression Profiling of Gliomas Strongly Predicts Survival, Cancer Research, 64: 6503-6510. 
}
\keyword{datasets}

