\name{glioma.x}
\alias{glioma.x}
\title{
Gene expression data of the patients with  gliomas
}
\description{
   This data set consists of gene expression for the 85 patients with gliomas.  
   Note that it contains a subset of the data published in Freije et al. (2004). 
}

\usage{data(glioma)}

\format{a matrix containing 100 genes and 85 samples}
\source{
Freije et al. (2004). Gene Expression Profiling of Gliomas Strongly Predicts Survival, Cancer Research, 64: 6503-6510. 
}
\keyword{datasets}

