\name{glioma.y}
\alias{glioma.y}
\title{
Survival data of the patients with  gliomas
}
\description{
This data set consists of survival for the 85 patients with gliomas.  
}

\usage{data(glioma)}


\format{a matrix containing survival and censoring status of 85 samples}
\source{
Freije et al. (2004). Gene Expression Profiling of Gliomas Strongly Predicts Survival, Cancer Research, 64: 6503-6510. 
}
\keyword{datasets}

